requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
boto3==1.37.20
botocore==1.37.20
Brotli==1.1.0
bs4==0.0.2
cdx_toolkit==0.9.37
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
daff==1.3.46
dbt==1.0.0.38.34
dbt-adapters==1.14.3
dbt-common==1.16.0
dbt-core==1.9.3
dbt-extractor==0.5.1
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
FastWARC==0.15.1
greenlet==3.1.1
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.6.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.4
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
protobuf==5.29.4
psycopg2-binary==2.9.10
py4j==0.10.9.7
pydantic==2.11.1
pydantic_core==2.33.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
Resiliparse==0.15.1
rpds-py==0.24.0
s3transfer==0.11.4
six==1.17.0
snowplow-tracker==1.1.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.3.0
warcio==1.7.5
wget==3.2
zipp==3.21.0
